fixed all kinds of small warnings from
* unused parameters
* unused variable
* compare between signed and unsigned

+++ trunk/src/chassis-plugin.c
@@ -86,7 +86,7 @@ GOptionEntry *chassis_plugin_get_options(chassis_plugin *p) {
 
 
 chassis_plugin* chassis_plugin_for_name(chassis *chas, gchar* plugin_name) {
-    int i;
+    gsize i;
 
     if (!chas || !plugin_name) return NULL;
